% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arit_prog.R
\name{arit.prog}
\alias{arit.prog}
\title{Progresión aritmética}
\usage{
arit.prog(x, n, size)
}
\arguments{
\item{x}{valor inicial de la progresión}

\item{n}{tamaño de la progresión}

\item{size}{tamaño del cambio}
}
\value{
Devuelve una progresión aritmética.
}
\description{
Función para crear progesiones aritméticas
}
\details{
Lo hicimos
}
\examples{
arit_prog(2,10,.5)
arit_prog(2,10,2)

}
