% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_prog.R
\name{geom_prog}
\alias{geom_prog}
\title{Progresión geométrica}
\usage{
geom_prog(x, n, size)
}
\arguments{
\item{x}{valor inicial de la progresión}

\item{n}{tamaño de la progresión}

\item{size}{tamaño del cambio}
}
\value{
Devuelve una progresión geométrica.
}
\description{
Función para crear progesiones geométricas
}
\details{
Estuvo dos tres fácil hacerla
}
\examples{
geom_prog(2,10,.5)
geom_prog(2,10,2)

}
